InsightSphere_server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
langchain-community
chromadb
pypdf
sentence-transformers
langchain-huggingface